src/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-storage-blob>=12.26.0",
    "boto3>=1.40.11",
    "fastapi[standard]>=0.116.1",
    "kafka-python>=2.2.15",
    "openai>=1.99.9",
    "pandas>=2.3.1",
    "psycopg>=3.2.9",
    "psycopg-binary>=3.2.9",
    "pycountry>=24.6.1",
    "pydantic-extra-types>=2.10.5",
    "redis>=6.4.0",
    "shortuuid>=1.0.13",
    "sqlalchemy>=2.0.42",
    "utils>=1.0.2",
    "uvicorn>=0.35.0",
]
